requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
distlib==0.3.9
Django==5.1.3
django-ckeditor==6.7.1
django-environ==0.11.2
django-js-asset==2.2.0
django-modeltranslation==0.19.10
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
filelock==3.16.1
idna==3.10
packaging==24.2
phonenumbers==8.13.49
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
PyJWT==2.9.0
pyTelegramBotAPI==4.23.0
python-gettext==5.0
requests==2.32.3
sqlparse==0.5.1
telebot==0.0.5
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.1
whitenoise==6.8.2